Reorganizando el codigo utilizando clase para tener todo mas organizado

+++ src/database/migrations/40-create-ventas.js
@@ -18,7 +18,7 @@ module.exports = {
         allowNull: false
       },
       cliente_id: {
-        type: Sequelize.INTEGER.UNSIGNED,
+        type: Sequelize.SMALLINT.UNSIGNED,
         allowNull: false,
         defaultValue: 1,
         references: {
